# api/openapi3/tasks.openapi.yaml
openapi: 3.0.0
info:
  title: Tasks
  version: 1.0.0
  description: |-
    Tasks — the HTTP contract of services/tasks.

    This file is the source of truth: the OpenAPI document (api/openapi3), the Go
    request/response types (libs/contracts/tasksapi) and the response validation
    in the tests are generated from it. Change the contract here, run
    `just contracts`, and the compiler tells you which handler no longer fits.

    Errors are RFC 9457 problem details (application/problem+json); every error
    response is a `Problem`, and business errors carry a stable `code`.
tags: []
paths:
  /tasks:
    post:
      operationId: TasksOps_create
      description: |-
        Create a task. The task and its `task.created` event (see events.tsp)
        are committed together (transactional outbox); the event reaches Kafka
        shortly after, even if the broker is down at the time of the request.

        With an Idempotency-Key, a retry of the same request returns the task the
        first one created (200, Idempotent-Replayed: true) instead of creating
        another; the same key with a different body is a 422. Keys are kept for
        24 hours.
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 255
      responses:
        '200':
          description: The request has succeeded.
          headers:
            ETag:
              required: true
              schema:
                type: string
            Idempotent-Replayed:
              required: true
              schema:
                type: string
                enum:
                  - 'true'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '201':
          description: The request has succeeded and a new resource has been created as a result.
          headers:
            ETag:
              required: true
              schema:
                type: string
            Location:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: A problem response with the given status and content type.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '422':
          description: A problem response with the given status and content type.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
    get:
      operationId: TasksOps_list
      description: |-
        List tasks, newest first, one page at a time (keyset pagination: pages
        do not shift when tasks are added or removed meanwhile).
      parameters:
        - name: limit
          in: query
          required: false
          description: Page size.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 200
            default: 50
          explode: false
        - name: cursor
          in: query
          required: false
          description: The next_cursor of the previous page.
          schema:
            type: string
          explode: false
      responses:
        '200':
          description: The request has succeeded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskList'
        '400':
          description: A problem response with the given status and content type.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
  /tasks/{id}:
    get:
      operationId: TasksOps_get
      description: 'Fetch one task; served from the cache when warm (X-Cache: hit|miss).'
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          headers:
            ETag:
              required: true
              schema:
                type: string
            X-Cache:
              required: false
              schema:
                type: string
                enum:
                  - hit
                  - miss
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: A problem response with the given status and content type.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
    patch:
      operationId: TasksOps_update
      description: |-
        Change the title and/or the done flag. With If-Match: "<version>" the
        change applies only if the task is still at that version — otherwise 412,
        re-read and retry. Emits `task.updated`.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: If-Match
          in: header
          required: false
          schema:
            type: string
      responses:
        '200':
          description: The request has succeeded.
          headers:
            ETag:
              required: true
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '400':
          description: A problem response with the given status and content type.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: A problem response with the given status and content type.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '412':
          description: A problem response with the given status and content type.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
    delete:
      operationId: TasksOps_delete
      description: |-
        Delete a task (and invalidate its cache entry); If-Match as for PATCH.
        Emits `task.deleted`.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: If-Match
          in: header
          required: false
          schema:
            type: string
      responses:
        '204':
          description: 'There is no content to send for this request, but the headers may be useful. '
        '400':
          description: A problem response with the given status and content type.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '404':
          description: A problem response with the given status and content type.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
        '412':
          description: A problem response with the given status and content type.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/Problem'
components:
  schemas:
    CreateTaskRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[^\u0000-\u0008\u000A-\u001F\u007F]*[^\s\u0000-\u001F\u007F][^\u0000-\u0008\u000A-\u001F\u007F]*$
          description: Title of the task; surrounding whitespace is trimmed (rules as Task.title).
      description: Body of POST /tasks.
    Problem:
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          description: URI reference identifying the problem type; "about:blank" when generic.
        title:
          type: string
          description: Short, human-readable summary of the problem type.
        status:
          type: integer
          format: int32
          minimum: 100
          maximum: 599
          description: HTTP status of this occurrence — always equals the response status.
        detail:
          type: string
          description: Human-readable explanation specific to this occurrence.
        instance:
          type: string
          description: URI reference identifying this occurrence (the request path).
        code:
          type: string
          description: Stable machine-readable code for business errors (e.g. task_not_found).
        request_id:
          type: string
          description: |-
            The request's correlation id — the same value as the X-Request-Id
            response header and the request_id field of the service's log lines.
      description: RFC 9457 problem details. Extension members are inlined (RFC 9457 §3.2).
    Task:
      type: object
      required:
        - id
        - title
        - done
        - version
        - created_at
      properties:
        id:
          type: string
          description: Server-assigned identifier (UUID).
        title:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[^\u0000-\u0008\u000A-\u001F\u007F]*[^\s\u0000-\u001F\u007F][^\u0000-\u0008\u000A-\u001F\u007F]*$
          description: |-
            Human-readable title, as stored: surrounding whitespace trimmed, 1–200
            characters, at least one of them not whitespace, no control characters
            other than tab (NUL in particular — PostgreSQL TEXT rejects it; found by
            Schemathesis). The pattern states the rules so that a generated,
            schema-compliant request is always accepted.
        done:
          type: boolean
          description: Completion flag.
        version:
          type: integer
          format: int64
          minimum: 1
          description: |-
            Starts at 1 and grows by one with every change. The ETag of the task is
            this number in quotes; send it back in If-Match to update or delete only
            the version you read.
        created_at:
          type: string
          format: date-time
          description: Creation time, RFC 3339 UTC, assigned by the database.
      description: A single to-do item.
    TaskList:
      type: object
      required:
        - tasks
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        next_cursor:
          type: string
          description: |-
            Opaque cursor of the next page — pass it as ?cursor= — absent on the
            last page.
      description: 'Body of GET /tasks: one page, newest first.'
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          pattern: ^[^\u0000-\u0008\u000A-\u001F\u007F]*[^\s\u0000-\u001F\u007F][^\u0000-\u0008\u000A-\u001F\u007F]*$
          description: New title; surrounding whitespace is trimmed (rules as Task.title).
        done:
          type: boolean
          description: New completion flag.
      description: 'Body of PATCH /tasks/{id}: the fields to change — at least one, or 400.'
servers:
  - url: http://localhost:8082
    description: local
    variables: {}