# libs/contracts/tasksapi/types.gen.go
// Package tasksapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package tasksapi

import (
	"time"
)

// CreateTaskRequest Body of POST /tasks.
type CreateTaskRequest struct {
	// Title Title of the task; surrounding whitespace is trimmed (rules as Task.title).
	Title string `json:"title"`
}

// Problem RFC 9457 problem details. Extension members are inlined (RFC 9457 §3.2).
type Problem struct {
	// Code Stable machine-readable code for business errors (e.g. task_not_found).
	Code *string `json:"code,omitempty"`

	// Detail Human-readable explanation specific to this occurrence.
	Detail *string `json:"detail,omitempty"`

	// Instance URI reference identifying this occurrence (the request path).
	Instance *string `json:"instance,omitempty"`

	// RequestId The request's correlation id — the same value as the X-Request-Id
	// response header and the request_id field of the service's log lines.
	RequestId *string `json:"request_id,omitempty"`

	// Status HTTP status of this occurrence — always equals the response status.
	Status int32 `json:"status"`

	// Title Short, human-readable summary of the problem type.
	Title string `json:"title"`

	// Type URI reference identifying the problem type; "about:blank" when generic.
	Type string `json:"type"`
}

// Task A single to-do item.
type Task struct {
	// CreatedAt Creation time, RFC 3339 UTC, assigned by the database.
	CreatedAt time.Time `json:"created_at"`

	// Done Completion flag.
	Done bool `json:"done"`

	// Id Server-assigned identifier (UUID).
	Id string `json:"id"`

	// Title Human-readable title, as stored: surrounding whitespace trimmed, 1–200
	// characters, at least one of them not whitespace, no control characters
	// other than tab (NUL in particular — PostgreSQL TEXT rejects it; found by
	// Schemathesis). The pattern states the rules so that a generated,
	// schema-compliant request is always accepted.
	Title string `json:"title"`

	// Version Starts at 1 and grows by one with every change. The ETag of the task is
	// this number in quotes; send it back in If-Match to update or delete only
	// the version you read.
	Version int64 `json:"version"`
}

// TaskList Body of GET /tasks: one page, newest first.
type TaskList struct {
	// NextCursor Opaque cursor of the next page — pass it as ?cursor= — absent on the
	// last page.
	NextCursor *string `json:"next_cursor,omitempty"`
	Tasks      []Task  `json:"tasks"`
}

// UpdateTaskRequest Body of PATCH /tasks/{id}: the fields to change — at least one, or 400.
type UpdateTaskRequest struct {
	// Done New completion flag.
	Done *bool `json:"done,omitempty"`

	// Title New title; surrounding whitespace is trimmed (rules as Task.title).
	Title *string `json:"title,omitempty"`
}

// TasksOpsListParams defines parameters for TasksOpsList.
type TasksOpsListParams struct {
	// Limit Page size.
	Limit *int32 `form:"limit,omitempty" json:"limit,omitempty"`

	// Cursor The next_cursor of the previous page.
	Cursor *string `form:"cursor,omitempty" json:"cursor,omitempty"`
}

// TasksOpsCreateParams defines parameters for TasksOpsCreate.
type TasksOpsCreateParams struct {
	IdempotencyKey *string `json:"Idempotency-Key,omitempty"`
}

// TasksOpsDeleteParams defines parameters for TasksOpsDelete.
type TasksOpsDeleteParams struct {
	IfMatch *string `json:"If-Match,omitempty"`
}

// TasksOpsUpdateParams defines parameters for TasksOpsUpdate.
type TasksOpsUpdateParams struct {
	IfMatch *string `json:"If-Match,omitempty"`
}

// TasksOpsCreateJSONRequestBody defines body for TasksOpsCreate for application/json ContentType.
type TasksOpsCreateJSONRequestBody = CreateTaskRequest

// TasksOpsUpdateJSONRequestBody defines body for TasksOpsUpdate for application/json ContentType.
type TasksOpsUpdateJSONRequestBody = UpdateTaskRequest